take 29

+++ Src/Xt.Synth0.Model/Param/ParamInfo.cs
@@ -27,7 +27,7 @@ namespace Xt.Synth0.Model
 			ParamType.Note => NoteNames[value],
 			ParamType.Toggle => value == 0 ? "Off" : "On",
 			ParamType.Type => ((UnitType)value).ToString(),
-			ParamType.Percent => ((value - Min) / (Max - Min)).ToString("P0"),
+			ParamType.Percent => (((double)value - Min) / (Max - Min)).ToString("P0"),
 			_ => throw new InvalidOperationException()
 		};
 	}
